baseline_methods/FinalWork/requirements.txt
# 核心数据处理
numpy>=1.21.0
pandas>=1.3.0

# 降维和聚类
umap-learn>=0.5.3
hdbscan>=0.8.29

# 机器学习
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 工具库
tqdm>=4.62.0

# 可选：GPU加速UMAP (需要CUDA)
# cuml>=21.10  # RAPIDS库，需要NVIDIA GPU
